feat(contract) removed context
