+++ components/layout/Providers.jsx
@@ -1,4 +1,3 @@
-import { ContractProvider } from "@/state/contract.context";
 import { Web3AuthProvider } from "@/state/web3auth.context";
 import { HumanProvider } from "@/state/human.context";
 
@@ -6,7 +5,7 @@ export const Providers = ({ children }) => {
   return (
     <Web3AuthProvider>
       <HumanProvider>
-        <ContractProvider>{children}</ContractProvider>
+        {children}
       </HumanProvider>
     </Web3AuthProvider>
   );
