--- utils/config.js
@@ -1,11 +1,0 @@
-export const config = {
-  CONTRACT: {
-    chainId: "0x13881",
-    address: "0x2CEDFf179BF88F7B4b1FFF9ca6d53393E956B74F",
-    abi: [
-      "function mint(address to, uint256 amount) public",
-      "function transfer(address to, uint256 amount) public returns (bool)",
-      "function decimals() public view returns (uint8)",
-    ],
-  },
-}
