+++ utils/index.js
@@ -1,4 +1,3 @@
 export * from "./api"
 export * from "./fetcherUtils"
 export * from "./queries"
-export * from "./config"
